+++ src/App.tsx
@@ -1,20 +1,23 @@
 import { Routes, Route } from 'react-router-dom';
-import { Layout } from '@/pages/Home/Layout';
 import { ROUTE_STRING } from '@/constants/routers';
-import OauthRedirectHandler from './components/OauthRedirectHandler';
-import Login from './pages/Auth/Login';
+import OauthRedirectHandler from './components/login/OauthRedirectHandler';
+import Login from './pages/Login';
+import Home from './pages/Home';
 
 const App = () => {
   return (
     <Routes>
       <Route path="/" element={<Login />} />
-      <Route path={ROUTE_STRING.OAUTH_REDIRECT} element={<OauthRedirectHandler />} />
+      <Route
+        path={ROUTE_STRING.OAUTH_REDIRECT}
+        element={<OauthRedirectHandler />}
+      />
       <Route path={ROUTE_STRING.LOGIN} element={<Login />} />
       <Route
         path={ROUTE_STRING.SIGNUP}
         element={<div>회원가입 페이지 입니다</div>}
       />
-      <Route path={ROUTE_STRING.HOME} element={<Layout />}>
+      <Route path={ROUTE_STRING.HOME} element={<Home />}>
         <Route
           path={ROUTE_STRING.SUBSCRIPTION}
           element={<div>구독 페이지 입니다</div>}
