--- src/pages/Home/Layout.tsx
@@ -1,11 +1,0 @@
-import { Outlet } from 'react-router-dom';
-import { NavigationBar } from '@/components/Footer';
-
-export const Layout = () => {
-  return (
-    <>
-      <Outlet />
-      <NavigationBar />
-    </>
-  );
-};
