--- src/types/auth.ts
@@ -1,1 +1,0 @@
-export type ProviderType = 'kakao' | 'naver' | 'google';
