+++ src/types/login.ts
@@ -1,0 +1,7 @@
+export type ProviderType = 'kakao' | 'naver' | 'google';
+
+export type LoginRequestDto = {
+  provider: ProviderType;
+  authorizationCode: string;
+  state?: string;
+};
