--- src/types/navigationBar/navigationBarTypes.ts
@@ -1,7 +1,0 @@
-export type NavigationButtonProps = {
-  navigationURL: string;
-  alt: string;
-  icon: string;
-  hoverIcon: string;
-  labelText: string;
-};
